requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==4.2.2
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
branca==0.4.2
bs4==0.0.1
cachetools==5.3.1
certifi==2022.12.7
cffi==1.16.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.7
cloudpickle==2.0.0
colorama==0.4.6
contourpy==1.1.1
cryptography==41.0.4
cycler==0.11.0
dataclasses-json==0.5.14
distlib==0.3.7
docopt==0.6.2
entrypoints==0.4
faiss-cpu==1.7.4
Faker==19.12.0
favicon==0.7.0
filelock==3.12.4
Flask==2.0.1
folium==0.12.1
fonttools==4.42.1
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
googletrans==3.1.0a0
greenlet==3.0.1
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
htbuilder==0.6.2
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langchain==0.0.154
lxml==4.9.3
Markdown==3.5
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
mdurl==0.1.2
more-itertools==10.1.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.7
numpy==1.26.0
openai==0.28.1
openapi-schema-pydantic==1.2.4
oscrypto==1.3.0
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
pipreqs==0.4.13
platformdirs==3.11.0
prompts==0.0.1
protobuf==3.20.3
pyarrow==13.0.0
pycparser==2.21
pycryptodomex==3.19.0
pydantic==1.10.13
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
pymdown-extensions==10.3.1
Pympler==1.0.1
pyOpenSSL==23.2.0
pyparsing==3.1.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzbar==0.1.9
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rfc3986==1.5.0
rich==13.6.0
rpds-py==0.10.6
semver==3.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
snowflake-connector-python==3.3.1
snowflake-snowpark-python==1.9.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.22
st-annotated-text==4.0.1
streamlit==1.18.1
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.3.4
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tenacity==8.2.3
tiktoken==0.5.1
tk==0.1.0
toml==0.10.2
tomlkit==0.12.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.15
validators==0.22.0
virtualenv==20.24.6
watchdog==3.0.0
Werkzeug==2.3.7
yarg==0.1.9
yarl==1.9.2
zipp==3.17.0